Move all plotting to plotting.py
